CRUD de usuarios

+++ src/main/java/curso/java/tienda/repository/UsuarioRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import curso.java.tienda.models.Usuarios;
 
-public interface UsuarioRepository extends JpaRepository<Usuarios, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuarios, Integer> {
 	Usuarios findById(int id);
 	Usuarios findByEmail(String email);
 }
